requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0

# Computer Vision & Object Detection
detectron2
timm>=0.9.0
opencv-python>=4.8.0

# NLP & Metrics
nltk>=3.8
pycocoevalcap>=1.2
rouge-score>=0.1.2
bert-score>=0.3.13

# Medical Imaging
pydicom>=2.4.0
SimpleITK>=2.3.0
scikit-image>=0.21.0

# Data & Utilities
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
h5py>=3.9.0

# Medical-specific
radgraph
chexbert

# Experiment Tracking
tensorboard>=2.13.0
wandb>=0.15.0

# Utils
pyyaml>=6.0
scikit-learn>=1.3.0
einops>=0.7.0
